Refactor to fix module load order
