+++ packages/asset-swapper/src/utils/market_operation_utils/constants.ts
@@ -1,35 +1,51 @@
+import { ChainId } from '@0x/contract-addresses';
 import { BigNumber } from '@0x/utils';
 
-import { DEFAULT_GAS_SCHEDULE } from '../../constants';
+import { BridgeContractAddresses } from '../../types';
 
 import { SourceFilters } from './source_filters';
-import { CurveFunctionSelectors, CurveInfo, ERC20BridgeSource, GetMarketOrdersOpts } from './types';
+import {
+    CurveFillData,
+    CurveFunctionSelectors,
+    CurveInfo,
+    DODOFillData,
+    ERC20BridgeSource,
+    FeeSchedule,
+    FillData,
+    GetMarketOrdersOpts,
+    MultiHopFillData,
+    SushiSwapFillData,
+    UniswapV2FillData,
+} from './types';
 
 // tslint:disable: custom-no-magic-numbers no-bitwise
 
 /**
  * Valid sources for market sell.
  */
-export const SELL_SOURCE_FILTER = new SourceFilters([
-    ERC20BridgeSource.Native,
-    ERC20BridgeSource.Uniswap,
-    ERC20BridgeSource.UniswapV2,
-    ERC20BridgeSource.Eth2Dai,
-    ERC20BridgeSource.Kyber,
-    ERC20BridgeSource.Curve,
-    ERC20BridgeSource.Balancer,
-    // Bancor is sampled off-chain, but this list should only include on-chain sources (used in ERC20BridgeSampler)
-    // ERC20BridgeSource.Bancor,
-    ERC20BridgeSource.MStable,
-    ERC20BridgeSource.Mooniswap,
-    ERC20BridgeSource.Swerve,
-    ERC20BridgeSource.SnowSwap,
-    ERC20BridgeSource.SushiSwap,
-    ERC20BridgeSource.Shell,
-    ERC20BridgeSource.MultiHop,
-    ERC20BridgeSource.Dodo,
-    ERC20BridgeSource.Cream,
-]);
+export const SELL_SOURCE_FILTER = new SourceFilters(
+    [
+        ERC20BridgeSource.Native,
+        ERC20BridgeSource.Uniswap,
+        ERC20BridgeSource.UniswapV2,
+        ERC20BridgeSource.Eth2Dai,
+        ERC20BridgeSource.Kyber,
+        ERC20BridgeSource.Curve,
+        ERC20BridgeSource.Balancer,
+        // Bancor is sampled off-chain, but this list should only include on-chain sources (used in ERC20BridgeSampler)
+        // ERC20BridgeSource.Bancor,
+        ERC20BridgeSource.MStable,
+        ERC20BridgeSource.Mooniswap,
+        ERC20BridgeSource.Swerve,
+        ERC20BridgeSource.SnowSwap,
+        ERC20BridgeSource.SushiSwap,
+        ERC20BridgeSource.Shell,
+        ERC20BridgeSource.MultiHop,
+        ERC20BridgeSource.Dodo,
+        ERC20BridgeSource.Cream,
+    ],
+    [ERC20BridgeSource.MultiBridge],
+);
 
 /**
  * Valid sources for market buy.
@@ -56,23 +72,6 @@ export const BUY_SOURCE_FILTER = new SourceFilters(
     ],
     [ERC20BridgeSource.MultiBridge],
 );
-
-export const DEFAULT_GET_MARKET_ORDERS_OPTS: GetMarketOrdersOpts = {
-    // tslint:disable-next-line: custom-no-magic-numbers
-    runLimit: 2 ** 15,
-    excludedSources: [],
-    excludedFeeSources: [],
-    includedSources: [],
-    bridgeSlippage: 0.005,
-    maxFallbackSlippage: 0.05,
-    numSamples: 13,
-    sampleDistributionBase: 1.05,
-    feeSchedule: DEFAULT_GAS_SCHEDULE,
-    gasSchedule: DEFAULT_GAS_SCHEDULE,
-    exchangeProxyOverhead: () => ZERO_AMOUNT,
-    allowFallback: true,
-    shouldGenerateQuoteReport: false,
-};
 
 /**
  * Sources to poll for ETH fee price estimates.
@@ -285,3 +284,139 @@ export const ONE_SECOND_MS = 1000;
 export const NULL_BYTES = '0x';
 export const NULL_ADDRESS = '0x0000000000000000000000000000000000000000';
 export const COMPARISON_PRICE_DECIMALS = 5;
+
+const EMPTY_BRIDGE_ADDRESSES: BridgeContractAddresses = {
+    uniswapBridge: NULL_ADDRESS,
+    uniswapV2Bridge: NULL_ADDRESS,
+    eth2DaiBridge: NULL_ADDRESS,
+    kyberBridge: NULL_ADDRESS,
+    curveBridge: NULL_ADDRESS,
+    multiBridge: NULL_ADDRESS,
+    balancerBridge: NULL_ADDRESS,
+    bancorBridge: NULL_ADDRESS,
+    mStableBridge: NULL_ADDRESS,
+    mooniswapBridge: NULL_ADDRESS,
+    sushiswapBridge: NULL_ADDRESS,
+    shellBridge: NULL_ADDRESS,
+    dodoBridge: NULL_ADDRESS,
+    creamBridge: NULL_ADDRESS,
+};
+
+export const BRIDGE_ADDRESSES_BY_CHAIN: { [chainId in ChainId]: BridgeContractAddresses } = {
+    [ChainId.Mainnet]: {
+        uniswapBridge: '0x36691c4f426eb8f42f150ebde43069a31cb080ad',
+        uniswapV2Bridge: '0xdcd6011f4c6b80e470d9487f5871a0cba7c93f48',
+        kyberBridge: '0xadd97271402590564ddd8ad23cb5317b1fb0fffb',
+        eth2DaiBridge: '0x991c745401d5b5e469b8c3e2cb02c748f08754f1',
+        curveBridge: '0x1796cd592d19e3bcd744fbb025bb61a6d8cb2c09',
+        multiBridge: '0xc03117a8c9bde203f70aa911cb64a7a0df5ba1e1',
+        balancerBridge: '0xfe01821ca163844203220cd08e4f2b2fb43ae4e4',
+        bancorBridge: '0x259897d9699553edbdf8538599242354e957fb94',
+        mStableBridge: '0x2bf04fcea05f0989a14d9afa37aa376baca6b2b3',
+        mooniswapBridge: '0x02b7eca484ad960fca3f7709e0b2ac81eec3069c',
+        sushiswapBridge: '0x47ed0262a0b688dcb836d254c6a2e96b6c48a9f5',
+        shellBridge: '0x21fb3862eed7911e0f8219a077247b849846728d',
+        dodoBridge: '0xe9da66965a9344aab2167e6813c03f043cc7a6ca',
+        creamBridge: '0xb9d4bf2c8dab828f4ffb656acdb6c2b497d44f25',
+    },
+    [ChainId.Kovan]: {
+        ...EMPTY_BRIDGE_ADDRESSES,
+        uniswapBridge: '0x0e85f89f29998df65402391478e5924700c0079d',
+        uniswapV2Bridge: '0x7b3530a635d099de0534dc27e46cd7c57578c3c8',
+        eth2DaiBridge: '0x2d47147429b474d2e4f83e658015858a1312ed5b',
+        kyberBridge: '0xaecfa25920f892b6eb496e1f6e84037f59da7f44',
+        curveBridge: '0x81c0ab53a7352d2e97f682a37cba44e54647eefb',
+        balancerBridge: '0x407b4128e9ecad8769b2332312a9f655cb9f5f3a',
+    },
+    [ChainId.Rinkeby]: EMPTY_BRIDGE_ADDRESSES,
+    [ChainId.Ropsten]: EMPTY_BRIDGE_ADDRESSES,
+    [ChainId.Ganache]: EMPTY_BRIDGE_ADDRESSES,
+};
+
+// tslint:disable:custom-no-magic-numbers
+export const DEFAULT_GAS_SCHEDULE: FeeSchedule = {
+    [ERC20BridgeSource.Native]: () => 150e3,
+    [ERC20BridgeSource.Uniswap]: () => 90e3,
+    [ERC20BridgeSource.LiquidityProvider]: () => 140e3,
+    [ERC20BridgeSource.Eth2Dai]: () => 400e3,
+    [ERC20BridgeSource.Kyber]: () => 500e3,
+    [ERC20BridgeSource.Curve]: fillData => {
+        const poolAddress = (fillData as CurveFillData).pool.poolAddress.toLowerCase();
+        switch (poolAddress) {
+            case '0xa5407eae9ba41422680e2e00537571bcc53efbfd':
+            case '0x93054188d876f558f4a66b2ef1d97d16edf0895b':
+            case '0x7fc77b5c7614e1533320ea6ddc2eb61fa00a9714':
+            case '0xbebc44782c7db0a1a60cb6fe97d0b483032ff1c7':
+                return 150e3;
+            case '0xa2b47e3d5c44877cca798226b7b8118f9bfb7a56':
+                return 750e3;
+            case '0x45f783cce6b7ff23b2ab2d70e416cdb7d6055f51':
+                return 850e3;
+            case '0x79a8c46dea5ada233abaffd40f3a0a2b1e5a4f27':
+                return 1e6;
+            case '0x52ea46506b9cc5ef470c5bf89f17dc28bb35d85c':
+                return 600e3;
+            default:
+                throw new Error(`Unrecognized Curve address: ${poolAddress}`);
+        }
+    },
+    [ERC20BridgeSource.MultiBridge]: () => 350e3,
+    [ERC20BridgeSource.UniswapV2]: (fillData?: FillData) => {
+        // TODO: Different base cost if to/from ETH.
+        let gas = 90e3;
+        const path = (fillData as UniswapV2FillData).tokenAddressPath;
+        if (path.length > 2) {
+            gas += (path.length - 2) * 60e3; // +60k for each hop.
+        }
+        return gas;
+    },
+    [ERC20BridgeSource.SushiSwap]: (fillData?: FillData) => {
+        // TODO: Different base cost if to/from ETH.
+        let gas = 95e3;
+        const path = (fillData as SushiSwapFillData).tokenAddressPath;
+        if (path.length > 2) {
+            gas += (path.length - 2) * 60e3; // +60k for each hop.
+        }
+        return gas;
+    },
+    [ERC20BridgeSource.Balancer]: () => 120e3,
+    [ERC20BridgeSource.Cream]: () => 300e3,
+    [ERC20BridgeSource.MStable]: () => 700e3,
+    [ERC20BridgeSource.Mooniswap]: () => 220e3,
+    [ERC20BridgeSource.Swerve]: () => 150e3,
+    [ERC20BridgeSource.Shell]: () => 300e3,
+    [ERC20BridgeSource.MultiHop]: (fillData?: FillData) => {
+        const firstHop = (fillData as MultiHopFillData).firstHopSource;
+        const secondHop = (fillData as MultiHopFillData).secondHopSource;
+        const firstHopGas = DEFAULT_GAS_SCHEDULE[firstHop.source]!(firstHop.fillData);
+        const secondHopGas = DEFAULT_GAS_SCHEDULE[secondHop.source]!(secondHop.fillData);
+        return new BigNumber(firstHopGas)
+            .plus(secondHopGas)
+            .plus(30e3)
+            .toNumber();
+    },
+    [ERC20BridgeSource.Dodo]: (fillData?: FillData) => {
+        const isSellBase = (fillData as DODOFillData).isSellBase;
+        // Sell base is cheaper as it is natively supported
+        // sell quote requires additional calculation and overhead
+        return isSellBase ? 440e3 : 540e3;
+    },
+};
+// tslint:enable:custom-no-magic-numbers
+
+export const DEFAULT_GET_MARKET_ORDERS_OPTS: GetMarketOrdersOpts = {
+    // tslint:disable-next-line: custom-no-magic-numbers
+    runLimit: 2 ** 15,
+    excludedSources: [],
+    excludedFeeSources: [],
+    includedSources: [],
+    bridgeSlippage: 0.005,
+    maxFallbackSlippage: 0.05,
+    numSamples: 13,
+    sampleDistributionBase: 1.05,
+    feeSchedule: DEFAULT_GAS_SCHEDULE,
+    gasSchedule: DEFAULT_GAS_SCHEDULE,
+    exchangeProxyOverhead: () => ZERO_AMOUNT,
+    allowFallback: true,
+    shouldGenerateQuoteReport: false,
+};
